+++ test/testing_data.dart
@@ -1,5 +1,3 @@
-part of sprintf_test;
-
 var expectedTestData = {
   '': {
     '%': ['|%|', '|%|', '|%|', '|%|', '|%|'],